[DAEMON-68] Finished up the initial implementation of appcd-subprocess and the SubprocessManager.
Upgraded to the latest and greatest gawk@4.0.2.

Fixed Dispatcher statuses for requests that return a Response object.

Major surgery on appcd-response. All instances of 'code' are now 'statusCode'. Lots of simplification and cleanup of the internal Message class. Custom error names are now rendering correctly in stack traces.

Lots of unit tests and improved code coverage.