Upgraded to gawk@4.1.0 to fix issue with the internal gawk state.
Fixed bug in Dispatcher so now you can pass in an empty string which is the equivalent to '/'.

Fixed bug with shutting down the PluginManager.

Fixed bug in SubprocessManager spawn handler where it would continue to write to the response even if the stream has been closed.

Fixed issues in test scripts.