appcd-plugin@1.1.1
Reset the plugin's 'stack' message when the plugin is stopped.

Fixed error handling when a plugin fails to activate or deactivate.

Added '/appcd/plugin/status/:name?/:version?' service to get a plugin's status without invoking the plugin.

Updated the plugin service docs with the status endpoint and fixed some typos.